stylify Home component
